data

+++ app.py
@@ -149,7 +149,7 @@ elif pagina_selecionada == "Consultas":
     df['Saída Tarde'] = pd.to_datetime(df['Saída Tarde'], format="%H:%M")
 
     # Converter timedelta para horas
-    df['Total trabalhado'] = df['Data'].astype(float)
+    df['Total trabalhado'] = df['Data']
 
     # Exibir o DataFrame na página
     st.write(df)
